+++ fabric/src/main/java/io/tebex/plugin/TebexPlugin.java
@@ -196,7 +196,7 @@ public class TebexPlugin implements Platform, DedicatedServerModInitializer {
     private Optional<ServerPlayerEntity> getPlayer(Object player) {
         if (player == null) return Optional.empty();
 
-        if (isOnlineMode()) {
+        if (isOnlineMode() && !isGeyser() && player instanceof UUID) {
             return Optional.ofNullable(server.getPlayerManager().getPlayer((UUID) player));
         }
 
@@ -217,6 +217,11 @@ public class TebexPlugin implements Platform, DedicatedServerModInitializer {
     @Override
     public String getVersion() {
         return MOD_VERSION;
+    }
+
+    @Override
+    public String getStoreType() {
+        return storeService.getStoreInformation() == null ? "" : storeService.getStoreInformation().getStore().getGameType();
     }
 
     @Override