+++ velocity/src/main/java/io/tebex/plugin/store/command/sub/DebugCommand.java
@@ -6,9 +6,10 @@ import io.tebex.plugin.TebexPlugin;
 import io.tebex.plugin.command.SubCommand;
 import io.tebex.sdk.platform.config.ProxyPlatformConfig;
 import io.tebex.sdk.util.StringUtil;
-import net.kyori.adventure.text.Component;
 
 import java.io.IOException;
+
+import static net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer.legacySection;
 
 public class DebugCommand extends SubCommand {
     public DebugCommand(TebexPlugin platform) {
@@ -23,21 +24,21 @@ public class DebugCommand extends SubCommand {
         YamlDocument configFile = config.getYamlDocument();
 
         if (StringUtil.isTruthy(args[0])) {
-            sender.sendMessage(Component.text("§b[Tebex] §7Debug mode enabled."));
+            sender.sendMessage(legacySection().deserialize("§b[Tebex] §7Debug mode enabled."));
             config.setVerbose(true);
             configFile.set("verbose", true);
         } else if (StringUtil.isFalsy(args[0])){
-            sender.sendMessage(Component.text("§b[Tebex] §7Debug mode disabled."));
+            sender.sendMessage(legacySection().deserialize("§b[Tebex] §7Debug mode disabled."));
             config.setVerbose(false);
             configFile.set("verbose", false);
         } else {
-            sender.sendMessage(Component.text("§b[Tebex] §7Invalid command usage. Use /tebex " + this.getName() + " " + getUsage()));
+            sender.sendMessage(legacySection().deserialize("§b[Tebex] §7Invalid command usage. Use /tebex " + this.getName() + " " + getUsage()));
         }
 
         try {
             configFile.save();
         } catch (IOException e) {
-            sender.sendMessage(Component.text("§b[Tebex] §7Failed to save configuration file."));
+            sender.sendMessage(legacySection().deserialize("§b[Tebex] §7Failed to save configuration file."));
         }
     }