+++ velocity/src/main/java/io/tebex/plugin/store/command/sub/ForceCheckCommand.java
@@ -3,7 +3,8 @@ package io.tebex.plugin.command.sub;
 import com.velocitypowered.api.command.CommandSource;
 import io.tebex.plugin.TebexPlugin;
 import io.tebex.plugin.command.SubCommand;
-import net.kyori.adventure.text.Component;
+
+import static net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer.legacySection;
 
 public class ForceCheckCommand extends SubCommand {
     private final TebexPlugin platform;
@@ -15,12 +16,12 @@ public class ForceCheckCommand extends SubCommand {
 
     @Override
     public void execute(CommandSource sender, String[] args) {
-        if(! platform.isStoreSetup()) {
-            sender.sendMessage(Component.text("§cTebex is not setup yet!"));
+        if(! platform.isSetup()) {
+            sender.sendMessage(legacySection().deserialize("§cTebex is not setup yet!"));
             return;
         }
 
-        sender.sendMessage(Component.text("§b[Tebex] §7Performing force check..."));
+        sender.sendMessage(legacySection().deserialize("§b[Tebex] §7Performing force check..."));
         getPlatform().performCheck(false);
     }